shouldConvertToUsdToEuro
GRÖNT!

+++ src/test/java/com/example/tdd/MoneyConverterTest.java
@@ -1,5 +1,6 @@
 package com.example.tdd;
 
+import org.example.MoneyConverter;
 import org.junit.jupiter.api.Test;
 import java.math.BigDecimal;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -10,7 +11,7 @@ public class MoneyConverterTest {
     void shouldConvertUsdToEuro() {
         BigDecimal usd = new BigDecimal("100.00");
 
-        BigDecimal result = MoneyConverter.usdToEuro(usd);
+        BigDecimal result = BigDecimal.valueOf(MoneyConverter.usdToEuro(usd));
 
         assertThat(result).isEqualByComparingTo("92.00");
     }
